+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/tools/item/ItemWorktableTool.java
@@ -1,7 +1,7 @@
 package com.codetaylor.mc.artisanworktables.modules.tools.item;
 
 import com.codetaylor.mc.artisanworktables.modules.tools.ModuleTools;
-import com.codetaylor.mc.artisanworktables.modules.tools.reference.EnumMaterial;
+import com.codetaylor.mc.athenaeum.reference.EnumMaterial;
 import com.codetaylor.mc.artisanworktables.modules.tools.reference.EnumWorktableToolType;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -40,6 +40,12 @@ public class ItemWorktableTool
   public EnumMaterial getMaterial() {
 
     return this.material;
+  }
+
+  @Override
+  public boolean isEnchantable(ItemStack stack) {
+
+    return false;
   }
 
   @Nonnull