+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/ModuleWorktables.java
@@ -1,6 +1,7 @@
 package com.codetaylor.mc.artisanworktables.modules.worktables;
 
 import com.codetaylor.mc.artisanworktables.ModArtisanWorktables;
+import com.codetaylor.mc.artisanworktables.modules.toolbox.block.BlockToolbox;
 import com.codetaylor.mc.artisanworktables.modules.worktables.block.BlockWorktable;
 import com.codetaylor.mc.artisanworktables.modules.worktables.item.ItemWorktable;
 import com.codetaylor.mc.artisanworktables.modules.worktables.network.SPacketWorktableTab;
@@ -98,10 +99,12 @@ public class ModuleWorktables
         TileEntityWorktableBasic.class,
         TileEntityWorktableBlacksmith.class,
         TileEntityWorktableCarpenter.class,
+        TileEntityWorktableChemist.class,
         TileEntityWorktableEngineer.class,
         TileEntityWorktableJeweler.class,
         TileEntityWorktableMage.class,
         TileEntityWorktableMason.class,
+        TileEntityWorktableScribe.class,
         TileEntityWorktableTailor.class
     );
   }